Improve an error message for missing source files.
When a source file is missing Mercury reports which other source file included
it, but not the whole chain of dependencies.  This is an experiment to report
the whole chain to see if it's more useful.

It doesn't appear to be useful.  Since it often includes the builtin module
and can get quite long.  The example here is a small one:

** Module `string' is imported or included by the chain of modules: `builtin' -> `private_builtin' -> `int' -> `table_builtin' -> `io' -> `char' -> `require'.

compiler/make.module_dep_file.m:
    As above.

compiler/make.m:
compiler/make.dependencies.m:
    Track the "stack" of dependencies as we calculate module dependencies.