Fix ML tailrecursion bug
Zoltan identified this bug where the code that checks for and warns about
tail recursion, and the code that implements self-TCO in the MLDS backend
have different conditions for what a tail call is.

This change makes the warning code stricter to bring it in line with the
actual TCO transformation.

WIP: this patch will probably become obsolete but I didn't want to forget
it, so adding it to my branch now.

compiler/ml_tailcall.m:
    Make the "at tail position" test stricter.

compiler/ml_optimize.m:
    We keep the test here, rather than relying only on the marking, since
    the no_return_call marking hasn't had the same test (that I know of).

    Use comments in both files to remind any developers of the now duplicate
    test.