wip
Zoltan this is the state of my wprkspace at the moment.

I think that this ml_gen_tscc makes sense.

ml_gen_tscc_procs will need a lot of work, there's a lot missing from it.
I was dithering on whether to rename variables at the same time as
generating MLDS code, or to rename the variables as a HLDS proc -> HLDS proc
transformation before generating the MLDS code.  I'm still not sure which is
best.